:recycle: Changed containers to position fixed
This is to resolve compatibility with certain mobile browsers that have a bottom search bar that minimises on scroll. Since we are disabling scroll on majority of our major components, in order to overcome the browser searchbar overlay we use position fixed, which happens to be relative to the visible area.